pyproject.toml
[project]
name = "urban-structure-analysis"
version = "0.1.0"
description = "都市構造解析システム - MFA/Lacunarity/Percolation統合解析"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.1.0",
    "pyproj>=3.6.0",
    "shapely>=2.0.0",
    "geopandas>=0.14.0",
    "rasterio>=1.3.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "joblib>=1.3.0",
    "opencv-python>=4.8.0",
    "dash>=2.14.0",
    "plotly>=5.18.0",
    "scipy>=1.11.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

